# Add RECORD_FUNCTION for aoti_xxx