Fixes #148650

add RECORD_FUNCTION for aoti_xxx

Example:
```python
import torch
from torch.testing._internal.common_utils import run_tests, TemporaryFileName, TestCase
from torch.utils import ThroughputBenchmark

from contextlib import nullcontext

from torch._dynamo import config
from torch._inductor import config as inductor_config


class TwoLayerNet(torch.jit.ScriptModule):
    def __init__(self, D_in, H, D_out):
        super().__init__()
        self.linear1 = torch.nn.Linear(D_in, H)
        self.linear2 = torch.nn.Linear(2 * H, D_out)

    @torch.jit.script_method
    def forward(self, x1, x2):
        h1_relu = self.linear1(x1).clamp(min=0)
        h2_relu = self.linear1(x2).clamp(min=0)
        cat = torch.cat((h1_relu, h2_relu), 1)
        y_pred = self.linear2(cat)
        return y_pred


class TwoLayerNetModule(torch.nn.Module):
    def __init__(self, D_in, H, D_out):
        super().__init__()
        self.linear1 = torch.nn.Linear(D_in, H)
        self.linear2 = torch.nn.Linear(2 * H, D_out)

    def forward(self, x1, x2):
        h1_relu = self.linear1(x1).clamp(min=0)
        h2_relu = self.linear1(x2).clamp(min=0)
        cat = torch.cat((h1_relu, h2_relu), 1)
        y_pred = self.linear2(cat)
        return y_pred

Module = TwoLayerNetModule
dtype = torch.bfloat16
config.error_on_recompile = True
inductor_config.cpp_wrapper = True
inductor_config.freezing = True
D_in = 10
H = 5
D_out = 15
B = 8

autocast = dtype != torch.float32
module = Module(D_in, H, D_out)

input = (torch.randn(B, D_in), torch.randn(B, D_in))

with torch.no_grad(), torch.amp.autocast("cpu", enabled=autocast, dtype=dtype):
    torch._dynamo.reset()
    module(*input)
    module = torch.compile(module)
    module(*input)
    module(*input)

with torch.autograd.profiler.profile() as prof:
    with torch.no_grad(), torch.amp.autocast("cpu", enabled=autocast, dtype=dtype):
        module(*input)
print(prof.key_averages().table(sort_by="self_cpu_time_total"))

```

Without this patch, profiler not show aoti_xxx:
------------------------------  ------------  ------------  ------------  ------------  ------------  ------------  
    Torch-Compiled Region: 0/0        80.52%     376.467us        82.69%     386.590us     386.590us             1  
      TorchDynamo Cache Lookup        17.31%      80.955us        17.31%      80.955us      80.955us             1  
                   aten::empty         2.17%      10.123us         2.17%      10.123us       3.374us             3  
------------------------------  ------------  ------------  ------------  ------------  ------------  ------------  
With this patch:
----------------------------------------------  ------------  ------------  ------------  ------------  ------------  ------------  
    at::native::mkldnn_linear_pointwise_binary        48.54%     269.358us        50.61%     280.830us      93.610us             3  
                    Torch-Compiled Region: 0/0        30.83%     171.110us        85.10%     472.230us     472.230us             1  
                      TorchDynamo Cache Lookup        14.90%      82.692us        14.90%      82.692us      82.692us             1  
                                   aten::empty         2.07%      11.472us         2.07%      11.472us       3.824us             3  
                      aoti_torch_empty_strided         1.63%       9.047us         1.63%       9.047us       4.524us             2  
               aoti_torch_delete_tensor_object         1.00%       5.558us         1.00%       5.558us       0.428us            13  
                aoti_torch__reinterpret_tensor         0.75%       4.175us         0.75%       4.175us       2.087us             2  
                       aoti_torch_get_data_ptr         0.27%       1.510us         0.27%       1.510us       0.151us            10  
----------------------------------------------  ------------  ------------  ------------  ------------  ------------  ------------

cc @voznesenskym @penguinwu @EikanWang @jgong5 @Guobing-Chen @XiaobingSuper @zhuhaozhe @blzheng @wenzhe-nrv @jiayisunx @ipiszy @chenyang78 @kadeng @muchulee8 @amjames @chauhang @aakhundov