@sanchitintel 2025-03-21
Hi, I had abandoned the same solution (@chunyuan-w's PoC) in #148290 because I wasn't sure if `RECORD_FUNCTION` should always be included, or if it should be enabled conditionally (which would require changes in the cpp-wrapper code).