@desertfire 2025-03-25
> Hi, I had abandoned the same solution (@chunyuan-w's PoC in #148290, which works for CPU, but I didn't make changes for other devices in that PoC) because I wasn't sure if `RECORD_FUNCTION` should always be added, or if it should be added conditionally (which would also require changes in the cpp-wrapper code).
> 
> Even for Inductor-CPU codegened functions, insertion of `RECORD_FUNCTION` calls is dependent upon an Inductor config.

I agree. We should insert `RECORD_FUNCTION` conditionally in the codegen phase.