@EikanWang 2025-03-25
@shiyang-weng , let's work out a design and prepare a mini RFC first.