@EikanWang 2025-03-28
@desertfire FYI, Shiyang and I had a short discussion days ago. We will follow your comments by adding `RECORD_FUNCTION` on demand. @shiyang-weng is working on this solution.